Divide MainForm.cs into several files
