+++ Wheel Tension Application/ParameterCalculations.cs
@@ -1,0 +1,39 @@
+﻿// This is an open source non-commercial project. Dear PVS-Studio, please check it.
+
+// PVS-Studio Static Code Analyzer for C, C++, C#, and Java: http://www.viva64.com
+
+using System.Collections.Generic;
+
+namespace Wheel_Tension_Application
+{
+    class ParameterCalculations
+    {
+        public List<float> CalculateSpokeAngles(List<float> tm1Reading)
+        {
+            var tm1ReadingLength = tm1Reading.Count;
+
+            float angle = 0;
+
+            List<float> angles = new List<float>(tm1ReadingLength);
+
+            float angleStep;
+            if (tm1Reading.Count > 0)
+            {
+                angleStep = (float)360 / tm1ReadingLength;
+            }
+            else
+            {
+                angleStep = (float)360 / 24;
+            }
+
+            for (int i = 0; i < tm1ReadingLength; i++)
+            {
+                angles.Add(angle);
+
+                angle += angleStep;
+            }
+
+            return angles;
+        }
+    }
+}
